setup golang implementation
